+++ packages/BoFilter/index.js
@@ -1,0 +1,9 @@
+import {
+    BoFilter
+} from './src/main';
+
+export default {
+    install(Vue) {
+        Vue.filter('boFilter', BoFilter)
+    }
+}